chore: remove index from blog components and update imports
