+++ src/app/ui/swap/SwapStepsExplainer.tsx
@@ -4,7 +4,7 @@ import type { ExtendedChain, Token } from '@lifi/sdk';
 import { Link as MuiLink, Typography, useColorScheme } from '@mui/material';
 import Link from 'next/link';
 import { Fragment } from 'react';
-import { Divider } from 'src/components/Blog';
+import { Divider } from 'src/components/Blog/BlogArticle/BlogArticle.style';
 import { DynamicPagesContainer } from 'src/components/DynamicPagesContainer';
 import StepDetail from 'src/components/StepDetail/StepDetail';
 import { getWidgetImageProps } from 'src/utils/image-generation/getWidgetImage';
