--- src/components/Blog/BlogArticle/index.ts
@@ -1,5 +1,0 @@
-export * from './BlogArticle';
-export * from './BlogArticle.style';
-export * from './ShareArticleIcon';
-export * from './ShareArticleIcon.style';
-export * from './ShareArticleIcons';
