--- src/components/Blog/BlogArticleCard/index.ts
@@ -1,2 +1,0 @@
-export * from './BlogArticleCard';
-export * from './BlogArticleCard.style';
