--- src/components/Blog/BlogCarousel/index.ts
@@ -1,1 +1,0 @@
-export * from './BlogCarousel';
