--- src/components/Blog/CTAs/BlogCTA/index.ts
@@ -1,2 +1,0 @@
-export * from './BlogCTA';
-export * from './BlogCTA.style';
