--- src/components/Blog/CTAs/InstructionsAccordion/index.ts
@@ -1,4 +1,0 @@
-export * from './InstructionsAccordion';
-export * from './InstructionsAccordion.style';
-export * from './InstructionsAccordionItem';
-export * from './InstructionsAccordionItem.style';
