--- src/components/Blog/CTAs/index.ts
@@ -1,2 +1,0 @@
-export * from './BlogCTA';
-export * from './InstructionsAccordion';
