--- src/components/Blog/FeaturedArticle/index.ts
@@ -1,3 +1,0 @@
-export * from './FeaturedArticle';
-export * from './FeaturedArticle.style';
-export * from './FeaturedArticleSkeleton';
