--- src/components/Blog/index.ts
@@ -1,7 +1,0 @@
-export * from './BlogArticle';
-export * from './BlogArticleCard';
-// export * from './BlogArticlesBoard';
-export * from './BlogCarousel';
-export * from './CTAs';
-export * from './CustomRichBlocks';
-export * from './FeaturedArticle';
